feat: 주문 내역 저장
- 주문 상태에 따라 주문 가능 여부 관리

+++ src/main/java/com/kr/economy/tradebatch/trade/application/OrderInCashCommand.java
@@ -17,8 +17,8 @@ public class OrderInCashCommand {
     private String ticker;
     private String trId;
     private OrderDvsnCode orderDvsnCode;
-    private Float sharePrice;
-    private Float orderPrice;
+    private int sharePrice;
+    private int orderPrice;
     private int qty;
     private KisOrderDvsnCode kisOrderDvsnCode;
 }
